In the ElectronicWfn object, a gradient attribute was added to help with minimizing molecules using qm packages. The gradient will be assigned to the wavefunction using the constructor.
A minimize module was added to the Psi4Potential. It calls the qsi4.opt function and saves the trajectory. Other appropriate changes were made to the Psi4Potential to facilitate execuation of this function.

A function was added to the Psi4Model to save a Psi4 optimization history as a MDT minimization trajectory. The the module was added outside the Psi4Potential class to allow for a user-defined Psi4 optimization history to be added as a minimization trajectory.

The Psi4interface module was also changed to allow for a user defined psi4history to be added.