The psi4_interface mdt_to_psi4 function now returns a psi4 molecule instead of a string.
The psi4_interface psi4_to_mdt function was written and now returns molecules.

Changes were made to the initiation of the multiplicity in the molecule class.

The Psi4 model was changed to accomodate charge and multiplicities from psi4 molecules. The interface was prepared to have a mdt wavefunction initiation function added later.